+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Forthcoming
 * Added subscription to /clock to display sim time on status bar.
 * Updated stamp format selection to preview the stamp format change.
 * Improved formatting of nodes, preferences, and details panels.
+* Improved search navigation.
 * Fixed exit to only require a single ctrl-c press.
 * Fixed message counts for nodes when loading messages from previous sessions.